Update scroll to top component
